backend/requirements.txt
fastapi
uvicorn
motor
httpx
pydantic
python-dotenv
